# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Organization feedback management system API",
        "title": "Kyooar API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "justdevelopitmx@proton.me"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/plans": {
            "get": {
                "description": "Retrieve all available subscription plans with their features and pricing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get available subscription plans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.SubscriptionPlan"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/can-create-organization": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if the authenticated user can create more organizations based on their subscription plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Check if user can create more organizations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.PermissionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/subscription": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the current subscription details for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get user's current subscription",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Subscription"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new subscription for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Create a new subscription",
                "parameters": [
                    {
                        "description": "Subscription details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Subscription"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel the current subscription for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Cancel user's subscription",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/subscription/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the current subscription usage details for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get user's current subscription usage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SubscriptionUsage"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateSubscriptionRequest": {
            "type": "object",
            "required": [
                "plan_id"
            ],
            "properties": {
                "plan_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deactivation_requested_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "subscription": {
                    "description": "Populated when needed"
                },
                "subscription_id": {
                    "type": "string"
                },
                "team_members": {
                    "description": "Organizations      []Organization  `json:\"organizations,omitempty\"`  // TODO: Add when organization domain is ready",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TeamMember"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.MemberRole": {
            "type": "string",
            "enum": [
                "OWNER",
                "ADMIN",
                "MANAGER",
                "VIEWER"
            ],
            "x-enum-varnames": [
                "RoleOwner",
                "RoleAdmin",
                "RoleManager",
                "RoleViewer"
            ]
        },
...
        "models.Subscription": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "account_id": {
                    "type": "string"
                },
                "cancel_at": {
                    "type": "string"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_period_end": {
                    "type": "string"
                },
                "current_period_start": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/models.SubscriptionPlan"
                },
                "plan_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.SubscriptionStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SubscriptionPlan": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_advanced_analytics": {
                    "type": "boolean"
                },
                "has_basic_analytics": {
                    "description": "Feature flags (as columns)",
                    "type": "boolean"
                },
                "has_custom_branding": {
                    "type": "boolean"
                },
                "has_feedback_explorer": {
                    "type": "boolean"
                },
                "has_priority_support": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "max_feedbacks_per_month": {
                    "type": "integer"
                },
                "max_organizations": {
                    "description": "Limits (as columns)",
                    "type": "integer"
                },
                "max_qr_codes": {
                    "type": "integer"
                },
                "max_team_members": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "trial_days": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SubscriptionStatus": {
            "type": "string",
            "enum": [
                "active",
                "pending",
                "canceled",
                "expired"
            ],
            "x-enum-varnames": [
                "SubscriptionActive",
                "SubscriptionPending",
                "SubscriptionCanceled",
                "SubscriptionExpired"
            ]
        },
        "models.SubscriptionUsage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "feedbacks_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "last_updated_at": {
                    "type": "string"
                },
                "locations_count": {
                    "type": "integer"
                },
                "organizations_count": {
                    "type": "integer"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "qr_codes_count": {
                    "type": "integer"
                },
                "subscription": {
                    "$ref": "#/definitions/models.Subscription"
                },
                "subscription_id": {
                    "type": "string"
                },
                "team_members_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TeamMember": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string"
                },
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "account_id": {
                    "description": "The organization account",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_at": {
                    "type": "string"
                },
                "invited_by": {
                    "type": "string"
                },
                "member": {
                    "$ref": "#/definitions/models.Account"
                },
                "member_id": {
                    "description": "The member's account ID",
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.MemberRole"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorData"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "services.PermissionResponse": {
            "type": "object",
            "properties": {
                "can_create": {
                    "type": "boolean"
                },
                "current_count": {
                    "type": "integer"
                },
                "max_allowed": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "subscription_status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}